+++ levels/snow/level4.txt
@@ -1,0 +1,10 @@
+0000000000
+0111111110
+0111111110
+0111111110
+0111111110
+0111111110
+0111111110
+0111111110
+0111111110
+0000000000
