+++ levels/snow/level7.txt
@@ -1,0 +1,10 @@
+0111111110
+0111111110
+0000110000
+0000110000
+0000110000
+0000110000
+0000110000
+0000110000
+0111111110
+0111111110
