+++ levels/snow/level8.txt
@@ -1,0 +1,6 @@
+110011
+110011
+000000
+100001
+010010
+001100
