+++ levels/snow/level9.txt
@@ -1,0 +1,10 @@
+0000000000
+0111111110
+0000000000
+0111111110
+0222222220
+0224444220
+0222222220
+0003333000
+0334444330
+0003333000
