+++ menu_seleccion.py
@@ -1,151 +1,133 @@
 import pygame
 from menu import Menu
-from utils import SettingsManager
-from enum import Enum
+from utils import BotonSeleccionNivel
+import ambiente as amb
+import os
+from mecanicas import Nivel
 
 # Evaluar si es conveniente fusionar este boton con el genérico
 # Ventajas de la fusión: Reutilización de código
 # Desventajas de la fusión: Aumento de complejidad, funcionalidades específicas
 
-class BotonSeleccionNivel:
-    def __init__(self, x, y, action=None, win=False):
-        self.x = x
-        self.y = y
+pygame.mixer.init()
+sonido_seleccion = pygame.mixer.Sound("resources/sounds/seleccion.wav")
+sonido_presion = pygame.mixer.Sound("resources/sounds/presion.wav")
 
-        # Carga de imagenes
-        self.imagen_boton_undone = pygame.transform.scale(pygame.image.load('resources/button/undone.png'), (100, 100))
-        self.imagen_boton_done = pygame.transform.scale(pygame.image.load('resources/button/done.png'), (100, 100))
-        self.imagen_boton_done_active = pygame.transform.scale(pygame.image.load('resources/button/done_active.png'), (100, 100))
-        self.imagen_boton_undone_active = pygame.transform.scale(pygame.image.load('resources/button/undone_active.png'), (100, 100))
 
-        self.imagen_actual = self.imagen_boton_undone
-        self.rect = self.imagen_boton_undone.get_rect(topleft=(x, y))
-        self.action = action
-        self.selected = False
-        self.win = win
+class MenuSeleccionNivel(Menu):
+    def __init__(self, menu_principal):
+        self.menu_principal = menu_principal
+        self.running = False
+        self.window = None
+        self.clock = None
+        self.niveles = []
+        self.botones_niveles = []
+        self.boton_seleccionado = 0
+        
 
-        # Animación del botón cuando se gana por primera vez
-        self.animacion_index = 0
-        self.animacion_tiempo = 0
-        self.animacion_win_completada = False
-        self.frames_boton_win = []
-        self.frame_index_boton_win = 0
-        for i in range(1,18):
-            img = pygame.image.load("resources/button/win_button/win_"+str(i)+".png")
-            img = pygame.transform.scale(img, (100, 100))
-            self.frames_boton_win.append(img)    
+        self.ambientes = {
+            amb.AmbienteEnum.INVIERNO: amb.Ambiente(amb.AmbienteEnum.INVIERNO, "resources/environments/images/snowG.png", "resources/environments/music/invierno.ogg", range(1,10))
+        }
 
-    def draw(self, screen):
-        if self.win and not self.animacion_win_completada:
-            self.win_animation()
-            self.imagen_actual = self.frames_boton_win[self.animacion_index]
-        elif self.win:
-            self.imagen_actual = self.imagen_boton_done_active if self.selected else self.imagen_boton_done
-        else:
-            self.imagen_actual = self.imagen_boton_undone_active if self.selected else self.imagen_boton_undone
-        screen.blit(self.imagen_actual, self.rect.topleft)
+        self.ambiente_actual = self.ambientes[amb.AmbienteEnum.INVIERNO]
+        self.pos_botones_ambiente_actual = self.ambiente_actual.get_pos_botones()
+        self.fondo_imagen = self.ambiente_actual.get_fondo()
+        self.ambiente_actual.cargar_musica()
 
-    def set_selected(self, selected):
-        self.selected = selected
+    def iniciar_pygame(self):
+        pygame.init()
+        self.window = pygame.display.set_mode((600, 600))
+        self.clock = pygame.time.Clock()
 
-    def check_click(self, pos):
-        if self.rect.collidepoint(pos):
-            if self.action:
-                self.action()
-            return True
-        return False
+        # obtener lista de archivos de la carpeta
+        carpeta_niveles = './levels/snow'
+        archivos_niveles = [f for f in os.listdir(carpeta_niveles) if os.path.isfile(os.path.join(carpeta_niveles, f))]
+
+        # Crear niveles
+        for archivo in archivos_niveles:
+            with open(os.path.join(carpeta_niveles, archivo), 'r') as file:
+                linea = 'a'
+                matriz = []
+                while(linea != ''):
+                    fila = []
+                    linea = file.readline()
+                    for char in linea:
+                        if(char.isdigit()):
+                            fila.append(int(char))
+
+                    if(len(fila) != 0):
+                        matriz.append(fila)
+                        
+                self.niveles.append(Nivel(matriz))
+
+        # Crear botones para los niveles
+        
+        for i, nivel in enumerate(self.niveles):
+            x, y = self.pos_botones_ambiente_actual[i]  # Obtener posición del nivel
+            boton = BotonSeleccionNivel(x, y, lambda n=nivel: self.iniciar_partida(n), nivel.isCompleted())
+            self.botones_niveles.append(boton)
+
+        #self.boton_volver = Boton("Volver", (10, 380), (200, 50), self.font, self.volver_al_menu_principal)
     
-    def check_hover(self, pos):
-        return self.rect.collidepoint(pos)
-    
-    def set_win_true(self):
-        self.win = True
+    def dibujar_menu(self):
+        fondo = pygame.transform.scale(self.ambiente_actual.get_fondo(), (600, 600))
+        self.window.blit(fondo, (0, 0))
+        
+        for boton in self.botones_niveles:
+            boton.draw(self.window)
+        #self.boton_volver.draw(self.window)
 
-    def win_animation(self):
-        self.animacion_tiempo += 1
-        if self.animacion_tiempo >= 5:
-            self.animacion_tiempo = 0
-            self.animacion_index += 1
-            if self.animacion_index >= len(self.frames_boton_win):
-                self.animacion_index = len(self.frames_boton_win) - 1
-                self.animacion_win_completada = True
-
-
-def main():
-    pygame.init()
-    screen = pygame.display.set_mode((800, 600))
-    clock = pygame.time.Clock()
-
-    botones = [
-        BotonSeleccionNivel(100, 100),
-        BotonSeleccionNivel(100, 200),
-        BotonSeleccionNivel(100, 300)
-    ]
-
-    selected_index = 0
-    botones[selected_index].set_selected(True)
-    nivel_completado= True
-    if nivel_completado:
-        botones[0].set_win_true()
-
-    running = True
-    while running:
-        for event in pygame.event.get():
-            if event.type == pygame.QUIT:
-                running = False
-            elif event.type == pygame.KEYDOWN:
-                if event.key == pygame.K_DOWN:
-                    botones[selected_index].set_selected(False)
-                    selected_index = (selected_index + 1) % len(botones)
-                    botones[selected_index].set_selected(True)
-                elif event.key == pygame.K_UP:
-                    botones[selected_index].set_selected(False)
-                    selected_index = (selected_index - 1) % len(botones)
-                    botones[selected_index].set_selected(True)
-            elif event.type == pygame.MOUSEMOTION:
-                for i, boton in enumerate(botones):
-                    if boton.check_hover(event.pos):
-                        botones[selected_index].set_selected(False)
-                        selected_index = i
-                        botones[selected_index].set_selected(True)
-                        break
-            elif event.type == pygame.MOUSEBUTTONDOWN:
-                if event.button == 1:  # Left mouse button
-                    for i, boton in enumerate(botones):
-                        if boton.check_click(event.pos):
-                            # Aquí puedes agregar la acción para entrar al nivel
-                            print(f"Botón {i} clickeado")
-                            break
-
-        screen.fill((0, 0, 0))
-        for boton in botones:
-            boton.draw(screen)
         pygame.display.flip()
-        clock.tick(60)
 
-    pygame.quit()
+    def iniciar_menu(self):
+        self.iniciar_pygame()
+        self.running = True
 
-class AmbienteEnum(Enum):
-    INVIERNO = "invierno"
-    MONTAÑA = "montaña"
-    PRADO = "prado"
+        while self.running:
+            self.clock.tick(144)
+            self.dibujar_menu()
 
-class Ambiente:
-    def __init__(self, tipo: AmbienteEnum, fondo, musica, niveles):
-        self.tipo = tipo
-        self.fondo = pygame.image.load(fondo)
-        self.musica = musica
-        self.niveles = niveles
+            for event in pygame.event.get():
+                if event.type == pygame.QUIT:
+                    self.running = False
+                    pygame.quit()
+                elif event.type == pygame.MOUSEMOTION:
+                    for i, boton in enumerate(self.botones_niveles):
+                        if boton.check_hover(event.pos):
+                            self.botones_niveles[self.boton_seleccionado].set_selected(False)
+                            self.boton_seleccionado = i
+                            self.botones_niveles[self.boton_seleccionado].set_selected(True)
+                elif event.type == pygame.KEYDOWN:
+                    if event.key == pygame.K_ESCAPE:
+                        self.volver_al_menu_principal()
+                    elif event.key == pygame.K_RIGHT:
+                        self.botones_niveles[self.boton_seleccionado].set_selected(False)
+                        self.boton_seleccionado = (self.boton_seleccionado + 1) % len(self.botones_niveles)
+                        self.botones_niveles[self.boton_seleccionado].set_selected(True)
+                        sonido_seleccion.play()
+                    elif event.key == pygame.K_LEFT:
+                        self.botones_niveles[self.boton_seleccionado].set_selected(False)
+                        self.boton_seleccionado = (self.boton_seleccionado - 1) % len(self.botones_niveles)
+                        self.botones_niveles[self.boton_seleccionado].set_selected(True)
+                        sonido_seleccion.play()
+                    elif event.key == pygame.K_RETURN:
+                        self.botones_niveles[self.boton_seleccionado].action()
+                        sonido_presion.play()
+                for boton in self.botones_niveles:
+                    boton.handle_event(event)
+            
+            self.clock.tick(10)
+        pygame.quit()
 
-    def cargar_musica(self):
-        pygame.mixer.music.load(self.musica)
-        pygame.mixer.music.play(-1)
+    def iniciar_partida(self, nivel):
+        self.cerrar_menu()
+        self.menu_principal.iniciar_partida(nivel)
 
-    def get_fondo(self):
-        return self.fondo
-
-    def get_niveles(self):
-        return self.niveles
+    def volver_al_menu_principal(self):
+        self.cerrar_menu()
+        self.menu_principal.iniciar_menu()
     
-if __name__ == "__main__":
-    main()
+    def cerrar_menu(self):
+        self.running = False
+        self.niveles = []
+        self.botones_niveles = []
